Move old_rails contents into new rails app
One file, the old_rails/license_builder.rb, was unused in the old
imazen_store repo, so I removed it entirely. Looks like it was maybe
intended to be a wrapper for ImazenLicensing::LicenseGenerator, but in
any event, it doesn't appear to be doing anything currently.

I also re-added actionview/railties because I'd forgotten we will need
views for mailers.