Add expires_on method
Right now this will only work for subscriptions cancelled after 3 years.

The perpetual addons are trickier because they aren't showing up under
the 'addons' key in chargebee data, and I'm only sometimes seeing them
under other keys, other times not at all. I added a has_perpetual_addon?
method to stub out and test that when it's true we make expires_on nil,
intending to fill out the implementation of has_perpetual_addon? when we
have a reliable way of checking for that.